backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
# 生产环境核心依赖（非开发工具类）
dependencies = [
    # FastAPI 核心依赖
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",

    # 数据库
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",

    # 异步支持
    "httpx>=0.25.0",

    # 数据验证和序列化
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",

    # 安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # 环境管理
    "python-dotenv>=1.0.0",

    # 文档
    "python-multipart>=0.0.6",

    # 生产环境
    "gunicorn>=21.2.0"
]

# 开发依赖分组（仅开发/测试时使用）
[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.27.0"
]